Fri, Apr 8, 2022, 3:46 PM +00:00

+++ readability/readability.c
@@ -95,6 +95,6 @@ int grade_calculator(letters, words, sentences)
     double S = (float)sentences / (float)words * 100;
     double index = 0.0588 * L - 0.296 * S - 15.8;
     printf("index before round: %f", index);
-    double round(double index);
+    double round(float index);
     return index;
 }
